agent-azure/requirements.txt
azure-ai-agents>=1.0.0
azure-ai-contentsafety>=1.0.0
azure-identity>=1.15.0
opentelemetry-sdk==1.39.0
azure-monitor-opentelemetry==1.8.6
python-dotenv>=0.21.0
pyyaml>=6.0
